limit_test.go
package increase_test

import (
	"context"
	"errors"
	"testing"

	"github.com/increase/increase-go"
	"github.com/increase/increase-go/option"
)

func TestLimitNewWithOptionalParams(t *testing.T) {
	c := increase.NewClient(option.WithAPIKey("APIKey"), option.WithBaseURL("http://127.0.0.1:4010"))
	_, err := c.Limits.New(context.TODO(), increase.LimitNewParams{Metric: increase.F(increase.LimitNewParamsMetricCount), Interval: increase.F(increase.LimitNewParamsIntervalTransaction), ModelID: increase.F("account_in71c4amph0vgo2qllky"), Value: increase.F(int64(1234))})
	if err != nil {
		var apierr *increase.Error
		if errors.As(err, &apierr) {
			println(apierr.DumpRequest(true))
		}
		t.Fatalf("err should be nil: %s", err.Error())
	}
}

func TestLimitGet(t *testing.T) {
	c := increase.NewClient(option.WithAPIKey("APIKey"), option.WithBaseURL("http://127.0.0.1:4010"))
	_, err := c.Limits.Get(
		context.TODO(),
		"limit_fku42k0qtc8ulsuas38q",
	)
	if err != nil {
		var apierr *increase.Error
		if errors.As(err, &apierr) {
			println(apierr.DumpRequest(true))
		}
		t.Fatalf("err should be nil: %s", err.Error())
	}
}

func TestLimitUpdate(t *testing.T) {
	c := increase.NewClient(option.WithAPIKey("APIKey"), option.WithBaseURL("http://127.0.0.1:4010"))
	_, err := c.Limits.Update(
		context.TODO(),
		"limit_fku42k0qtc8ulsuas38q",
		increase.LimitUpdateParams{Status: increase.F(increase.LimitUpdateParamsStatusInactive)},
	)
	if err != nil {
		var apierr *increase.Error
		if errors.As(err, &apierr) {
			println(apierr.DumpRequest(true))
		}
		t.Fatalf("err should be nil: %s", err.Error())
	}
}

func TestLimitListWithOptionalParams(t *testing.T) {
	c := increase.NewClient(option.WithAPIKey("APIKey"), option.WithBaseURL("http://127.0.0.1:4010"))
	_, err := c.Limits.List(context.TODO(), increase.LimitListParams{Cursor: increase.F("string"), Limit: increase.F(int64(0)), ModelID: increase.F("x"), Status: increase.F("x")})
	if err != nil {
		var apierr *increase.Error
		if errors.As(err, &apierr) {
			println(apierr.DumpRequest(true))
		}
		t.Fatalf("err should be nil: %s", err.Error())
	}
}
